backend_py/requirements.txt
# Web framework
fastapi==0.114.2
uvicorn[standard]==0.30.6

# Config and validation
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1

# Supabase client
supabase==2.9.0

# Google Gemini
google-generativeai==0.8.2

# HTTP
httpx==0.27.2

# PDF processing
PyPDF2==3.0.1
python-multipart==0.0.6

# Tavily API for search
tavily-python==0.3.3

# Logging
loguru==0.7.2

# File handling
aiofiles==23.2.0

# JSON handling
orjson==3.9.10
